Metrics units as YAML comments (#166)
* refactor roc_curves_models.py to use pymatviz.roc_curve_plotly

* use ruamel.yaml to add units as EoL comments in metrics.geo_opt

* reorder landing page

* make write_discovery_metrics_to_yaml add units as EoL comments as well

* move (scripts -> matbench_discovery)/metrics-which-is-better.yml

* install pymatviz from main
