+++ .github/workflows/test.yml
@@ -36,7 +36,7 @@ jobs:
       fail-fast: false
       matrix:
         script:
-          - scripts/metrics/discovery_tables.py
+          - scripts/metrics/eval_discovery.py
           - scripts/model_figs/run_all.py
     steps:
       - name: Check out repository
