+++ models/alignn/alignn.yml
@@ -51,62 +51,59 @@ metrics:
     pred_file: models/alignn/2023-06-02-alignn-wbm-IS2RE.csv.gz
     pred_col: e_form_per_atom_alignn
     full_test_set:
-      F1: 0.565
-      DAF: 2.921
-      Precision: 0.501
-      Recall: 0.649
-      Accuracy: 0.829
-      TPR: 0.649
-      FPR: 0.134
-      TNR: 0.866
-      FNR: 0.351
-      TP: 28598.0
-      FP: 28464.0
-      TN: 184407.0
-      FN: 15494.0
-      MAE: 0.092
-      RMSE: 0.154
-      R2: 0.274
-      missing_preds: 1
-      missing_percent: 0.00%
+      F1: 0.565 # fraction
+      DAF: 2.921 # dimensionless
+      Precision: 0.501 # fraction
+      Recall: 0.649 # fraction
+      Accuracy: 0.829 # fraction
+      TPR: 0.649 # fraction
+      FPR: 0.134 # fraction
+      TNR: 0.866 # fraction
+      FNR: 0.351 # fraction
+      TP: 28598.0 # count
+      FP: 28464.0 # count
+      TN: 184407.0 # count
+      FN: 15494.0 # count
+      MAE: 0.092 # eV/atom
+      RMSE: 0.154 # eV/atom
+      R2: 0.274 # dimensionless
+      missing_preds: 1 # count
+      missing_percent: 0.00% # fraction
     most_stable_10k:
-      F1: 0.748
-      DAF: 3.905
-      Precision: 0.597
-      Recall: 1.0
-      Accuracy: 0.597
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 5969.0
-      FP: 4031.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.176
-      RMSE: 0.247
-      R2: 0.081
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.748 # fraction
+      DAF: 3.905 # dimensionless
+      Precision: 0.597 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.597 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 5969.0 # count
+      FP: 4031.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.176 # eV/atom
+      RMSE: 0.247 # eV/atom
+      R2: 0.081 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.567
-      DAF: 3.206
-      Precision: 0.49
-      Recall: 0.672
-      Accuracy: 0.841
-      TPR: 0.672
-      FPR: 0.128
-      TNR: 0.872
-      FNR: 0.328
-      TP: 22436.0
-      FP: 23346.0
-      TN: 158768.0
-      FN: 10938.0
-      MAE: 0.093
-      RMSE: 0.154
-      R2: 0.297
-      missing_preds: 0
-      missing_percent: 0.00%
-# model pre-trained by NIST (not shown in MBD since performed worse)
-# training_set:
-#   title: MP Graphs
+      F1: 0.567 # fraction
+      DAF: 3.206 # dimensionless
+      Precision: 0.49 # fraction
+      Recall: 0.672 # fraction
+      Accuracy: 0.841 # fraction
+      TPR: 0.672 # fraction
+      FPR: 0.128 # fraction
+      TNR: 0.872 # fraction
+      FNR: 0.328 # fraction
+      TP: 22436.0 # count
+      FP: 23346.0 # count
+      TN: 158768.0 # count
+      FN: 10938.0 # count
+      MAE: 0.093 # eV/atom
+      RMSE: 0.154 # eV/atom
+      R2: 0.297 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
