+++ models/cgcnn/cgcnn.yml
@@ -48,59 +48,59 @@ metrics:
     pred_file: models/cgcnn/2023-01-26-cgcnn-ens=10-wbm-IS2RE.csv.gz
     pred_col: e_form_per_atom_mp2020_corrected_pred_ens
     full_test_set:
-      F1: 0.51
-      DAF: 2.631
-      Precision: 0.451
-      Recall: 0.587
-      Accuracy: 0.807
-      TPR: 0.587
-      FPR: 0.148
-      TNR: 0.852
-      FNR: 0.413
-      TP: 25895.0
-      FP: 31474.0
-      TN: 181397.0
-      FN: 18197.0
-      MAE: 0.135
-      RMSE: 0.229
-      R2: -0.624
-      missing_preds: 4
-      missing_percent: 0.00%
+      F1: 0.51 # fraction
+      DAF: 2.631 # dimensionless
+      Precision: 0.451 # fraction
+      Recall: 0.587 # fraction
+      Accuracy: 0.807 # fraction
+      TPR: 0.587 # fraction
+      FPR: 0.148 # fraction
+      TNR: 0.852 # fraction
+      FNR: 0.413 # fraction
+      TP: 25895.0 # count
+      FP: 31474.0 # count
+      TN: 181397.0 # count
+      FN: 18197.0 # count
+      MAE: 0.135 # eV/atom
+      RMSE: 0.229 # eV/atom
+      R2: -0.624 # dimensionless
+      missing_preds: 4 # count
+      missing_percent: 0.00% # fraction
     most_stable_10k:
-      F1: 0.745
-      DAF: 3.88
-      Precision: 0.593
-      Recall: 1.0
-      Accuracy: 0.593
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 5931.0
-      FP: 4069.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.165
-      RMSE: 0.23
-      R2: 0.181
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.745 # fraction
+      DAF: 3.88 # dimensionless
+      Precision: 0.593 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.593 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 5931.0 # count
+      FP: 4069.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.165 # eV/atom
+      RMSE: 0.23 # eV/atom
+      R2: 0.181 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.507
-      DAF: 2.855
-      Precision: 0.436
-      Recall: 0.605
-      Accuracy: 0.818
-      TPR: 0.605
-      FPR: 0.143
-      TNR: 0.857
-      FNR: 0.395
-      TP: 20191.0
-      FP: 26073.0
-      TN: 156041.0
-      FN: 13183.0
-      MAE: 0.138
-      RMSE: 0.233
-      R2: -0.603
-      missing_preds: 2
-      missing_percent: 0.00%
+      F1: 0.507 # fraction
+      DAF: 2.855 # dimensionless
+      Precision: 0.436 # fraction
+      Recall: 0.605 # fraction
+      Accuracy: 0.818 # fraction
+      TPR: 0.605 # fraction
+      FPR: 0.143 # fraction
+      TNR: 0.857 # fraction
+      FNR: 0.395 # fraction
+      TP: 20191.0 # count
+      FP: 26073.0 # count
+      TN: 156041.0 # count
+      FN: 13183.0 # count
+      MAE: 0.138 # eV/atom
+      RMSE: 0.233 # eV/atom
+      R2: -0.603 # dimensionless
+      missing_preds: 2 # count
+      missing_percent: 0.00% # fraction
