+++ models/chgnet/chgnet.yml
@@ -67,68 +67,69 @@ metrics:
   geo_opt:
     pred_file: models/chgnet/2023-12-21-chgnet-0.3.0-wbm-IS2RE.json.gz
     pred_col: chgnet_structure
-    rmsd: 0.0216
-    symmetry_decrease: 0.2526
-    symmetry_match: 0.5833
-    symmetry_increase: 0.1525
-    n_structs: 250779.0
+    rmsd: 0.0216 # Å
+    symmetry_decrease: 0.2526 # fraction
+    symmetry_match: 0.5833 # fraction
+    symmetry_increase: 0.1525 # fraction
+    n_sym_ops_mae: 3.2731
+    n_structures: 250779.0 # count
   discovery:
     pred_file: models/chgnet/2023-12-21-chgnet-0.3.0-wbm-IS2RE.csv.gz
     pred_col: e_form_per_atom_chgnet
     full_test_set:
-      F1: 0.612
-      DAF: 3.038
-      Precision: 0.521
-      Recall: 0.74
-      Accuracy: 0.839
-      TPR: 0.74
-      FPR: 0.141
-      TNR: 0.859
-      FNR: 0.26
-      TP: 32642.0
-      FP: 29979.0
-      TN: 182892.0
-      FN: 11450.0
-      MAE: 0.061
-      RMSE: 0.1
-      R2: 0.69
-      missing_preds: 2
-      missing_percent: 0.00%
+      F1: 0.612 # fraction
+      DAF: 3.038 # dimensionless
+      Precision: 0.521 # fraction
+      Recall: 0.74 # fraction
+      Accuracy: 0.839 # fraction
+      TPR: 0.74 # fraction
+      FPR: 0.141 # fraction
+      TNR: 0.859 # fraction
+      FNR: 0.26 # fraction
+      TP: 32642.0 # count
+      FP: 29979.0 # count
+      TN: 182892.0 # count
+      FN: 11450.0 # count
+      MAE: 0.061 # eV/atom
+      RMSE: 0.1 # eV/atom
+      R2: 0.69 # dimensionless
+      missing_preds: 2 # count
+      missing_percent: 0.00% # fraction
     most_stable_10k:
-      F1: 0.92
-      DAF: 5.567
-      Precision: 0.851
-      Recall: 1.0
-      Accuracy: 0.851
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 8511.0
-      FP: 1489.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.063
-      RMSE: 0.095
-      R2: 0.816
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.92 # fraction
+      DAF: 5.567 # dimensionless
+      Precision: 0.851 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.851 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 8511.0 # count
+      FP: 1489.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.063 # eV/atom
+      RMSE: 0.095 # eV/atom
+      R2: 0.816 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.613
-      DAF: 3.361
-      Precision: 0.514
-      Recall: 0.758
-      Accuracy: 0.851
-      TPR: 0.758
-      FPR: 0.132
-      TNR: 0.868
-      FNR: 0.242
-      TP: 25313.0
-      FP: 23955.0
-      TN: 158159.0
-      FN: 8061.0
-      MAE: 0.063
-      RMSE: 0.103
-      R2: 0.689
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.613 # fraction
+      DAF: 3.361 # dimensionless
+      Precision: 0.514 # fraction
+      Recall: 0.758 # fraction
+      Accuracy: 0.851 # fraction
+      TPR: 0.758 # fraction
+      FPR: 0.132 # fraction
+      TNR: 0.868 # fraction
+      FNR: 0.242 # fraction
+      TP: 25313.0 # count
+      FP: 23955.0 # count
+      TN: 158159.0 # count
+      FN: 8061.0 # count
+      MAE: 0.063 # eV/atom
+      RMSE: 0.103 # eV/atom
+      R2: 0.689 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
