+++ models/eqV2/eqV2-m-omat-mp-salex.yml
@@ -87,63 +87,69 @@ metrics:
   geo_opt:
     pred_file: models/eqV2/eqV2-86M-omat-salex-mp.json.gz
     pred_col: eqV2-86M-omat-mp-salex_structure
+    rmsd: 0.0138 # Å
+    n_sym_ops_mae: 10.0558
+    symmetry_decrease: 0.8611 # fraction
+    symmetry_match: 0.1382 # fraction
+    symmetry_increase: 0.0005 # fraction
+    n_structures: 256614.0 # count
   discovery:
     pred_file: models/eqV2/eqV2-m-omat-mp-salex.csv.gz
     pred_col: e_form_per_atom_eqV2-86M-omat-mp-salex
     full_test_set:
-      F1: 0.896
-      DAF: 5.243
-      Precision: 0.9
-      Recall: 0.893
-      Accuracy: 0.965
-      TPR: 0.893
-      FPR: 0.021
-      TNR: 0.979
-      FNR: 0.107
-      TP: 39379.0
-      FP: 4393.0
-      TN: 208478.0
-      FN: 4713.0
-      MAE: 0.02
-      RMSE: 0.071
-      R2: 0.842
-      missing_preds: 2
-      missing_percent: 0.00%
+      F1: 0.896 # fraction
+      DAF: 5.243 # dimensionless
+      Precision: 0.9 # fraction
+      Recall: 0.893 # fraction
+      Accuracy: 0.965 # fraction
+      TPR: 0.893 # fraction
+      FPR: 0.021 # fraction
+      TNR: 0.979 # fraction
+      FNR: 0.107 # fraction
+      TP: 39379.0 # count
+      FP: 4393.0 # count
+      TN: 208478.0 # count
+      FN: 4713.0 # count
+      MAE: 0.02 # eV/atom
+      RMSE: 0.071 # eV/atom
+      R2: 0.842 # dimensionless
+      missing_preds: 2 # count
+      missing_percent: 0.00% # fraction
     most_stable_10k:
-      F1: 0.988
-      DAF: 6.382
-      Precision: 0.976
-      Recall: 1.0
-      Accuracy: 0.976
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 9756.0
-      FP: 244.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.015
-      RMSE: 0.066
-      R2: 0.904
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.988 # fraction
+      DAF: 6.382 # dimensionless
+      Precision: 0.976 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.976 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 9756.0 # count
+      FP: 244.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.015 # eV/atom
+      RMSE: 0.066 # eV/atom
+      R2: 0.904 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.917
-      DAF: 6.047
-      Precision: 0.924
-      Recall: 0.91
-      Accuracy: 0.975
-      TPR: 0.91
-      FPR: 0.014
-      TNR: 0.986
-      FNR: 0.09
-      TP: 30372.0
-      FP: 2481.0
-      TN: 179633.0
-      FN: 3002.0
-      MAE: 0.02
-      RMSE: 0.072
-      R2: 0.848
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.917 # fraction
+      DAF: 6.047 # dimensionless
+      Precision: 0.924 # fraction
+      Recall: 0.91 # fraction
+      Accuracy: 0.975 # fraction
+      TPR: 0.91 # fraction
+      FPR: 0.014 # fraction
+      TNR: 0.986 # fraction
+      FNR: 0.09 # fraction
+      TP: 30372.0 # count
+      FP: 2481.0 # count
+      TN: 179633.0 # count
+      FN: 3002.0 # count
+      MAE: 0.02 # eV/atom
+      RMSE: 0.072 # eV/atom
+      R2: 0.848 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
