+++ models/mace/mace.yml
@@ -72,68 +72,69 @@ metrics:
   geo_opt:
     pred_file: models/mace/2023-12-11-mace-wbm-IS2RE-FIRE.json.gz
     pred_col: mace_structure
-    rmsd: 0.0194
-    symmetry_decrease: 0.035
-    symmetry_match: 0.7361
-    symmetry_increase: 0.2243
-    n_structs: 243070.0
+    rmsd: 0.0194 # Å
+    symmetry_decrease: 0.035 # fraction
+    symmetry_match: 0.7361 # fraction
+    symmetry_increase: 0.2243 # fraction
+    n_sym_ops_mae: 1.8584
+    n_structures: 243070.0 # count
   discovery:
     pred_file: models/mace/2023-12-11-mace-wbm-IS2RE-FIRE.csv.gz
     pred_col: e_form_per_atom_mace
     full_test_set:
-      F1: 0.668
-      DAF: 3.4
-      Precision: 0.583
-      Recall: 0.781
-      Accuracy: 0.867
-      TPR: 0.781
-      FPR: 0.115
-      TNR: 0.885
-      FNR: 0.219
-      TP: 34420.0
-      FP: 24576.0
-      TN: 188295.0
-      FN: 9672.0
-      MAE: 0.055
-      RMSE: 0.099
-      R2: 0.698
-      missing_preds: 38
-      missing_percent: 0.01%
+      F1: 0.668 # fraction
+      DAF: 3.4 # dimensionless
+      Precision: 0.583 # fraction
+      Recall: 0.781 # fraction
+      Accuracy: 0.867 # fraction
+      TPR: 0.781 # fraction
+      FPR: 0.115 # fraction
+      TNR: 0.885 # fraction
+      FNR: 0.219 # fraction
+      TP: 34420.0 # count
+      FP: 24576.0 # count
+      TN: 188295.0 # count
+      FN: 9672.0 # count
+      MAE: 0.055 # eV/atom
+      RMSE: 0.099 # eV/atom
+      R2: 0.698 # dimensionless
+      missing_preds: 38 # count
+      missing_percent: 0.01% # fraction
     most_stable_10k:
-      F1: 0.888
-      DAF: 5.221
-      Precision: 0.798
-      Recall: 1.0
-      Accuracy: 0.798
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 7982.0
-      FP: 2018.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.087
-      RMSE: 0.165
-      R2: 0.508
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.888 # fraction
+      DAF: 5.221 # dimensionless
+      Precision: 0.798 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.798 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 7982.0 # count
+      FP: 2018.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.087 # eV/atom
+      RMSE: 0.165 # eV/atom
+      R2: 0.508 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.669
-      DAF: 3.777
-      Precision: 0.577
-      Recall: 0.796
-      Accuracy: 0.878
-      TPR: 0.796
-      FPR: 0.107
-      TNR: 0.893
-      FNR: 0.204
-      TP: 26582.0
-      FP: 19457.0
-      TN: 162657.0
-      FN: 6792.0
-      MAE: 0.057
-      RMSE: 0.101
-      R2: 0.697
-      missing_preds: 34
-      missing_percent: 0.02%
+      F1: 0.669 # fraction
+      DAF: 3.777 # dimensionless
+      Precision: 0.577 # fraction
+      Recall: 0.796 # fraction
+      Accuracy: 0.878 # fraction
+      TPR: 0.796 # fraction
+      FPR: 0.107 # fraction
+      TNR: 0.893 # fraction
+      FNR: 0.204 # fraction
+      TP: 26582.0 # count
+      FP: 19457.0 # count
+      TN: 162657.0 # count
+      FN: 6792.0 # count
+      MAE: 0.057 # eV/atom
+      RMSE: 0.101 # eV/atom
+      R2: 0.697 # dimensionless
+      missing_preds: 34 # count
+      missing_percent: 0.02% # fraction
