+++ models/mattersim/mattersim.yml
@@ -124,59 +124,59 @@ metrics:
     pred_file: models/mattersim/2024-06-16-mattersim-wbm-IS2RE.csv.gz
     pred_col: e_form_per_atom_mattersim
     full_test_set:
-      F1: 0.832
-      DAF: 4.838
-      Precision: 0.83
-      Recall: 0.834
-      Accuracy: 0.942
-      TPR: 0.834
-      FPR: 0.035
-      TNR: 0.965
-      FNR: 0.166
-      TP: 36754.0
-      FP: 7516.0
-      TN: 205355.0
-      FN: 7338.0
-      MAE: 0.026
-      RMSE: 0.079
-      R2: 0.809
-      missing_preds: 2
-      missing_percent: 0.00%
+      F1: 0.832 # fraction
+      DAF: 4.838 # dimensionless
+      Precision: 0.83 # fraction
+      Recall: 0.834 # fraction
+      Accuracy: 0.942 # fraction
+      TPR: 0.834 # fraction
+      FPR: 0.035 # fraction
+      TNR: 0.965 # fraction
+      FNR: 0.166 # fraction
+      TP: 36754.0 # count
+      FP: 7516.0 # count
+      TN: 205355.0 # count
+      FN: 7338.0 # count
+      MAE: 0.026 # eV/atom
+      RMSE: 0.079 # eV/atom
+      R2: 0.809 # dimensionless
+      missing_preds: 2 # count
+      missing_percent: 0.00% # fraction
     most_stable_10k:
-      F1: 0.989
-      DAF: 6.401
-      Precision: 0.978
-      Recall: 1.0
-      Accuracy: 0.978
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 9785.0
-      FP: 215.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.02
-      RMSE: 0.065
-      R2: 0.906
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.989 # fraction
+      DAF: 6.401 # dimensionless
+      Precision: 0.978 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.978 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 9785.0 # count
+      FP: 215.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.02 # eV/atom
+      RMSE: 0.065 # eV/atom
+      R2: 0.906 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.859
-      DAF: 5.646
-      Precision: 0.863
-      Recall: 0.856
-      Accuracy: 0.957
-      TPR: 0.856
-      FPR: 0.025
-      TNR: 0.975
-      FNR: 0.144
-      TP: 28562.0
-      FP: 4531.0
-      TN: 177583.0
-      FN: 4812.0
-      MAE: 0.026
-      RMSE: 0.08
-      R2: 0.812
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.859 # fraction
+      DAF: 5.646 # dimensionless
+      Precision: 0.863 # fraction
+      Recall: 0.856 # fraction
+      Accuracy: 0.957 # fraction
+      TPR: 0.856 # fraction
+      FPR: 0.025 # fraction
+      TNR: 0.975 # fraction
+      FNR: 0.144 # fraction
+      TP: 28562.0 # count
+      FP: 4531.0 # count
+      TN: 177583.0 # count
+      FN: 4812.0 # count
+      MAE: 0.026 # eV/atom
+      RMSE: 0.08 # eV/atom
+      R2: 0.812 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
