+++ models/voronoi_rf/voronoi-rf.yml
@@ -48,59 +48,59 @@ metrics:
     pred_file: models/voronoi_rf/2022-11-27-train-test/e-form-preds-IS2RE.csv.gz
     pred_col: e_form_per_atom_voronoi_rf
     full_test_set:
-      F1: 0.344
-      DAF: 1.509
-      Precision: 0.259
-      Recall: 0.511
-      Accuracy: 0.665
-      TPR: 0.511
-      FPR: 0.303
-      TNR: 0.697
-      FNR: 0.489
-      TP: 22517.0
-      FP: 64431.0
-      TN: 148440.0
-      FN: 21575.0
-      MAE: 0.141
-      RMSE: 0.206
-      R2: -0.316
-      missing_preds: 19
-      missing_percent: 0.01%
+      F1: 0.344 # fraction
+      DAF: 1.509 # dimensionless
+      Precision: 0.259 # fraction
+      Recall: 0.511 # fraction
+      Accuracy: 0.665 # fraction
+      TPR: 0.511 # fraction
+      FPR: 0.303 # fraction
+      TNR: 0.697 # fraction
+      FNR: 0.489 # fraction
+      TP: 22517.0 # count
+      FP: 64431.0 # count
+      TN: 148440.0 # count
+      FN: 21575.0 # count
+      MAE: 0.141 # eV/atom
+      RMSE: 0.206 # eV/atom
+      R2: -0.316 # dimensionless
+      missing_preds: 19 # count
+      missing_percent: 0.01% # fraction
     most_stable_10k:
-      F1: 0.551
-      DAF: 2.487
-      Precision: 0.38
-      Recall: 1.0
-      Accuracy: 0.38
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 3802.0
-      FP: 6198.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.349
-      RMSE: 0.417
-      R2: -1.012
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.551 # fraction
+      DAF: 2.487 # dimensionless
+      Precision: 0.38 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.38 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 3802.0 # count
+      FP: 6198.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.349 # eV/atom
+      RMSE: 0.417 # eV/atom
+      R2: -1.012 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.333
-      DAF: 1.579
-      Precision: 0.241
-      Recall: 0.535
-      Accuracy: 0.668
-      TPR: 0.535
-      FPR: 0.308
-      TNR: 0.692
-      FNR: 0.465
-      TP: 17854.0
-      FP: 56122.0
-      TN: 125992.0
-      FN: 15520.0
-      MAE: 0.148
-      RMSE: 0.212
-      R2: -0.329
-      missing_preds: 2
-      missing_percent: 0.00%
+      F1: 0.333 # fraction
+      DAF: 1.579 # dimensionless
+      Precision: 0.241 # fraction
+      Recall: 0.535 # fraction
+      Accuracy: 0.668 # fraction
+      TPR: 0.535 # fraction
+      FPR: 0.308 # fraction
+      TNR: 0.692 # fraction
+      FNR: 0.465 # fraction
+      TP: 17854.0 # count
+      FP: 56122.0 # count
+      TN: 125992.0 # count
+      FN: 15520.0 # count
+      MAE: 0.148 # eV/atom
+      RMSE: 0.212 # eV/atom
+      R2: -0.329 # dimensionless
+      missing_preds: 2 # count
+      missing_percent: 0.00% # fraction
