+++ models/wrenformer/wrenformer.yml
@@ -53,59 +53,59 @@ metrics:
     pred_file: models/wrenformer/2022-11-15-wrenformer-ens=10-IS2RE-preds.csv.gz
     pred_col: e_form_per_atom_wrenformer_pred_ens
     full_test_set:
-      F1: 0.479
-      DAF: 2.13
-      Precision: 0.365
-      Recall: 0.693
-      Accuracy: 0.741
-      TPR: 0.693
-      FPR: 0.249
-      TNR: 0.751
-      FNR: 0.307
-      TP: 30566.0
-      FP: 53076.0
-      TN: 159795.0
-      FN: 13526.0
-      MAE: 0.105
-      RMSE: 0.182
-      R2: -0.02
-      missing_preds: 7
-      missing_percent: 0.00%
+      F1: 0.479 # fraction
+      DAF: 2.13 # dimensionless
+      Precision: 0.365 # fraction
+      Recall: 0.693 # fraction
+      Accuracy: 0.741 # fraction
+      TPR: 0.693 # fraction
+      FPR: 0.249 # fraction
+      TNR: 0.751 # fraction
+      FNR: 0.307 # fraction
+      TP: 30566.0 # count
+      FP: 53076.0 # count
+      TN: 159795.0 # count
+      FN: 13526.0 # count
+      MAE: 0.105 # eV/atom
+      RMSE: 0.182 # eV/atom
+      R2: -0.02 # dimensionless
+      missing_preds: 7 # count
+      missing_percent: 0.00% # fraction
     most_stable_10k:
-      F1: 0.721
-      DAF: 3.691
-      Precision: 0.564
-      Recall: 1.0
-      Accuracy: 0.564
-      TPR: 1.0
-      FPR: 1.0
-      TNR: 0.0
-      FNR: 0.0
-      TP: 5642.0
-      FP: 4358.0
-      TN: 0.0
-      FN: 0.0
-      MAE: 0.182
-      RMSE: 0.239
-      R2: 0.138
-      missing_preds: 0
-      missing_percent: 0.00%
+      F1: 0.721 # fraction
+      DAF: 3.691 # dimensionless
+      Precision: 0.564 # fraction
+      Recall: 1.0 # fraction
+      Accuracy: 0.564 # fraction
+      TPR: 1.0 # fraction
+      FPR: 1.0 # fraction
+      TNR: 0.0 # fraction
+      FNR: 0.0 # fraction
+      TP: 5642.0 # count
+      FP: 4358.0 # count
+      TN: 0.0 # count
+      FN: 0.0 # count
+      MAE: 0.182 # eV/atom
+      RMSE: 0.239 # eV/atom
+      R2: 0.138 # dimensionless
+      missing_preds: 0 # count
+      missing_percent: 0.00% # fraction
     unique_prototypes:
-      F1: 0.466
-      DAF: 2.256
-      Precision: 0.345
-      Recall: 0.719
-      Accuracy: 0.745
-      TPR: 0.719
-      FPR: 0.25
-      TNR: 0.75
-      FNR: 0.281
-      TP: 23992.0
-      FP: 45575.0
-      TN: 136539.0
-      FN: 9382.0
-      MAE: 0.11
-      RMSE: 0.186
-      R2: -0.018
-      missing_preds: 5
-      missing_percent: 0.00%
+      F1: 0.466 # fraction
+      DAF: 2.256 # dimensionless
+      Precision: 0.345 # fraction
+      Recall: 0.719 # fraction
+      Accuracy: 0.745 # fraction
+      TPR: 0.719 # fraction
+      FPR: 0.25 # fraction
+      TNR: 0.75 # fraction
+      FNR: 0.281 # fraction
+      TP: 23992.0 # count
+      FP: 45575.0 # count
+      TN: 136539.0 # count
+      FN: 9382.0 # count
+      MAE: 0.11 # eV/atom
+      RMSE: 0.186 # eV/atom
+      R2: -0.018 # dimensionless
+      missing_preds: 5 # count
+      missing_percent: 0.00% # fraction
