+++ pyproject.toml
@@ -31,7 +31,7 @@ dependencies = [
   "pandas>=2.2.2",
   "plotly>=5.23",
   "pymatgen>=2024.8.9",
-  "pymatviz[export-figs,df-pdf-export,df-svg-export]>=0.14",
+  "pymatviz[export-figs,df-pdf-export,df-svg-export] @ git+https://github.com/janosh/pymatviz",
   "ruamel.yaml>=0.18.6",
   "scikit-learn>=1.5",
   "scipy>=1.13",
