--- scripts/metrics-which-is-better.yml
@@ -1,26 +1,0 @@
-discovery:
-  higher_is_better:
-    - Acc
-    - Accuracy
-    - AUC
-    - DAF
-    - F1
-    - Prec
-    - Precision
-    - R2
-    - R<sup>2</sup> # need both R2 and HTML version for different parts of the code
-    - Recall
-    - TN
-    - TNR
-    - TP
-    - TPR
-  lower_is_better: [MAE, RMSE, FPR, FNR, FP, FN]
-
-geo_opt:
-  higher_is_better: [Match]
-  # unclear if lower symmetry increase is really better, could be model actually found a higher symmetry lower-energy structure than DFT optimizer
-  lower_is_better: [RMSD, Decrease, Increase]
-
-phonons:
-  lower_is_better: [κ_SRME, κ<sub>SRME</sub>]
-  higher_is_better: []
