+++ site/src/routes/api/+page.svelte
@@ -7,13 +7,13 @@
 <style>
   /* select all but first module h1s */
   :global(h1[id^='module-']:not(:nth-of-type(2))) {
-    margin: 2em 0 0;
+    margin-top: 2em;
   }
   :global(h1 > kbd) {
     font-size: 25pt;
   }
   :global(hr) {
     border: 1px solid gray;
-    margin: 0 0 2em;
+    margin-bottom: 2em;
   }
 </style>
