+++ site/src/routes/models/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { ModelStatLabel, ModelStats } from '$lib'
   import { model_is_compliant, MODEL_METADATA, ModelCard } from '$lib'
-  import { discovery } from '$root/scripts/metrics-which-is-better.yml'
+  import { discovery } from '$pkg/metrics-which-is-better.yml'
   import Icon from '@iconify/svelte'
   import { interpolateCividis as cividis } from 'd3-scale-chromatic'
   import { ColorBar } from 'elementari'
@@ -81,7 +81,7 @@
   }
 </script>
 
-<div style="display: grid; margin: 1vw 3vw;">
+<div style="display: grid;">
   <h1>Leaderboard</h1>
 
   <p style="text-align: center;">
